Add custom officing residence
As the web service response does not include the date_of_birth in its response,
we use the interface's date. We still check if the given date is above the
minimum age.

As the Santiago de Compostela web service already checks that the given date
of birth is correct, we can use it to check if the user has the minimum allowed
age to participate.

Tests based on CensusApi were skipped (as this fork does not use it),
duplicated and adapted to pass when using the Santiago remote census response
format, so the officing residence features still work even without the date_of_birth
present on web service responses.